+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Window manager persistence moved to settings database
 - Routing persistence moved to settings database
 - Last read UID persistence moved to settings database
+- Improved logging
 
 ## [0.8.255] - 2023-02-02
 ### Changed: